+++ CHANGES.md
@@ -4,6 +4,7 @@
 * Added: Section/Grid/Col background images and archive/featured images now use <picture> and <source> elements for much better mobile speed performance by loading smaller images on smaller window sizes.
 * Added: Now add a context of "hidden" (without quotes) to any Section in the Sections template to disable it from being displayed on the front end.
 * Added: New `mai_scroll_to_elements` filter to add elements by class/id/etc to allow slow scroll to anchor links.
+* Added: New `mai_flex_entry_info` and `mai_flex_entry_meta` filter to specifically filter info/meta in [grid],
 * Changed: No longer remove entry-header element when banner area is displayed.
 * Changed: Updated WooCommerce templates to latest version.
 * Changed: Better "starts with" JS, without regex.